split up filesystem classes
so we have one class per file, and break them up into separate packages to make it easier to manage
