+++ h2/src/main/org/h2/store/fs/mem/FilePathMemLZF.java
@@ -1,0 +1,25 @@
+package org.h2.store.fs.mem;
+
+/**
+ * A memory file system that compresses blocks to conserve memory.
+ */
+class FilePathMemLZF extends FilePathMem {
+
+    @Override
+    public FilePathMem getPath(String path) {
+        FilePathMemLZF p = new FilePathMemLZF();
+        p.name = getCanonicalPath(path);
+        return p;
+    }
+
+    @Override
+    boolean compressed() {
+        return true;
+    }
+
+    @Override
+    public String getScheme() {
+        return "memLZF";
+    }
+
+}
