+++ h2/src/main/org/h2/store/fs/niomapped/FileNioMapped.java
@@ -1,9 +1,4 @@
-/*
- * Copyright 2004-2019 H2 Group. Multiple-Licensed under the MPL 2.0,
- * and the EPL 1.0 (https://h2database.com/html/license.html).
- * Initial Developer: H2 Group
- */
-package org.h2.store.fs;
+package org.h2.store.fs.niomapped;
 
 import java.io.EOFException;
 import java.io.IOException;
@@ -14,29 +9,13 @@ import java.nio.MappedByteBuffer;
 import java.nio.channels.FileChannel;
 import java.nio.channels.FileLock;
 import java.nio.channels.NonWritableChannelException;
+import java.nio.channels.FileChannel.MapMode;
 import java.nio.file.Paths;
 import java.util.concurrent.TimeUnit;
-
 import org.h2.engine.SysProperties;
+import org.h2.store.fs.FileBase;
+import org.h2.store.fs.FileUtils;
 import org.h2.util.MemoryUnmapper;
-
-/**
- * This file system stores files on disk and uses java.nio to access the files.
- * This class used memory mapped files.
- */
-public class FilePathNioMapped extends FilePathWrapper {
-
-    @Override
-    public FileChannel open(String mode) throws IOException {
-        return new FileNioMapped(name.substring(getScheme().length() + 1), mode);
-    }
-
-    @Override
-    public String getScheme() {
-        return "nioMapped";
-    }
-
-}
 
 /**
  * Uses memory mapped files.
