+++ h2/src/main/org/h2/store/fs/niomem/FileNioMemData.java
@@ -1,410 +1,15 @@
-/*
- * Copyright 2004-2019 H2 Group. Multiple-Licensed under the MPL 2.0,
- * and the EPL 1.0 (https://h2database.com/html/license.html).
- * Initial Developer: H2 Group
- */
-package org.h2.store.fs;
+package org.h2.store.fs.niomem;
 
-import java.io.IOException;
 import java.nio.ByteBuffer;
-import java.nio.channels.ClosedChannelException;
-import java.nio.channels.FileChannel;
-import java.nio.channels.FileLock;
 import java.nio.channels.NonWritableChannelException;
-import java.util.ArrayList;
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.TreeMap;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
-import org.h2.api.ErrorCode;
 import org.h2.compress.CompressLZF;
-import org.h2.message.DbException;
+import org.h2.store.fs.niomem.FileNioMemData.CompressItem;
+import org.h2.store.fs.niomem.FileNioMemData.CompressLaterCache;
 import org.h2.util.MathUtils;
-
-/**
- * This file system keeps files fully in memory. There is an option to compress
- * file blocks to save memory.
- */
-public class FilePathNioMem extends FilePath {
-
-    private static final TreeMap<String, FileNioMemData> MEMORY_FILES =
-            new TreeMap<>();
-
-    /**
-     * The percentage of uncompressed (cached) entries.
-     */
-    float compressLaterCachePercent = 1;
-
-    @Override
-    public FilePathNioMem getPath(String path) {
-        FilePathNioMem p = new FilePathNioMem();
-        p.name = getCanonicalPath(path);
-        return p;
-    }
-
-    @Override
-    public long size() {
-        return getMemoryFile().length();
-    }
-
-    @Override
-    public void moveTo(FilePath newName, boolean atomicReplace) {
-        synchronized (MEMORY_FILES) {
-            if (!atomicReplace && !name.equals(newName.name) &&
-                    MEMORY_FILES.containsKey(newName.name)) {
-                throw DbException.get(ErrorCode.FILE_RENAME_FAILED_2, name, newName + " (exists)");
-            }
-            FileNioMemData f = getMemoryFile();
-            f.setName(newName.name);
-            MEMORY_FILES.remove(name);
-            MEMORY_FILES.put(newName.name, f);
-        }
-    }
-
-    @Override
-    public boolean createFile() {
-        synchronized (MEMORY_FILES) {
-            if (exists()) {
-                return false;
-            }
-            getMemoryFile();
-        }
-        return true;
-    }
-
-    @Override
-    public boolean exists() {
-        if (isRoot()) {
-            return true;
-        }
-        synchronized (MEMORY_FILES) {
-            return MEMORY_FILES.get(name) != null;
-        }
-    }
-
-    @Override
-    public void delete() {
-        if (isRoot()) {
-            return;
-        }
-        synchronized (MEMORY_FILES) {
-            MEMORY_FILES.remove(name);
-        }
-    }
-
-    @Override
-    public List<FilePath> newDirectoryStream() {
-        ArrayList<FilePath> list = new ArrayList<>();
-        synchronized (MEMORY_FILES) {
-            for (String n : MEMORY_FILES.tailMap(name).keySet()) {
-                if (n.startsWith(name)) {
-                    list.add(getPath(n));
-                } else {
-                    break;
-                }
-            }
-            return list;
-        }
-    }
-
-    @Override
-    public boolean setReadOnly() {
-        return getMemoryFile().setReadOnly();
-    }
-
-    @Override
-    public boolean canWrite() {
-        return getMemoryFile().canWrite();
-    }
-
-    @Override
-    public FilePathNioMem getParent() {
-        int idx = name.lastIndexOf('/');
-        return idx < 0 ? null : getPath(name.substring(0, idx));
-    }
-
-    @Override
-    public boolean isDirectory() {
-        if (isRoot()) {
-            return true;
-        }
-        // TODO in memory file system currently
-        // does not really support directories
-        synchronized (MEMORY_FILES) {
-            return MEMORY_FILES.get(name) == null;
-        }
-    }
-
-    @Override
-    public boolean isAbsolute() {
-        // TODO relative files are not supported
-        return true;
-    }
-
-    @Override
-    public FilePathNioMem toRealPath() {
-        return this;
-    }
-
-    @Override
-    public long lastModified() {
-        return getMemoryFile().getLastModified();
-    }
-
-    @Override
-    public void createDirectory() {
-        if (exists() && isDirectory()) {
-            throw DbException.get(ErrorCode.FILE_CREATION_FAILED_1,
-                    name + " (a file with this name already exists)");
-        }
-        // TODO directories are not really supported
-    }
-
-    @Override
-    public FileChannel open(String mode) {
-        FileNioMemData obj = getMemoryFile();
-        return new FileNioMem(obj, "r".equals(mode));
-    }
-
-    private FileNioMemData getMemoryFile() {
-        synchronized (MEMORY_FILES) {
-            FileNioMemData m = MEMORY_FILES.get(name);
-            if (m == null) {
-                m = new FileNioMemData(name, compressed(), compressLaterCachePercent);
-                MEMORY_FILES.put(name, m);
-            }
-            return m;
-        }
-    }
-
-    protected boolean isRoot() {
-        return name.equals(getScheme() + ":");
-    }
-
-    /**
-     * Get the canonical path of a file (with backslashes replaced with forward
-     * slashes).
-     *
-     * @param fileName the file name
-     * @return the canonical path
-     */
-    protected static String getCanonicalPath(String fileName) {
-        fileName = fileName.replace('\\', '/');
-        int idx = fileName.lastIndexOf(':') + 1;
-        if (fileName.length() > idx && fileName.charAt(idx) != '/') {
-            fileName = fileName.substring(0, idx) + "/" + fileName.substring(idx);
-        }
-        return fileName;
-    }
-
-    @Override
-    public String getScheme() {
-        return "nioMemFS";
-    }
-
-    /**
-     * Whether the file should be compressed.
-     *
-     * @return true if it should be compressed.
-     */
-    boolean compressed() {
-        return false;
-    }
-
-}
-
-/**
- * A memory file system that compresses blocks to conserve memory.
- */
-class FilePathNioMemLZF extends FilePathNioMem {
-
-    @Override
-    boolean compressed() {
-        return true;
-    }
-
-    @Override
-    public FilePathNioMem getPath(String path) {
-        if (!path.startsWith(getScheme())) {
-            throw new IllegalArgumentException(path +
-                    " doesn't start with " + getScheme());
-        }
-        int idx1 = path.indexOf(':');
-        int idx2 = path.lastIndexOf(':');
-        final FilePathNioMemLZF p = new FilePathNioMemLZF();
-        if (idx1 != -1 && idx1 != idx2) {
-            p.compressLaterCachePercent = Float.parseFloat(path.substring(idx1 + 1, idx2));
-        }
-        p.name = getCanonicalPath(path);
-        return p;
-    }
-
-    @Override
-    protected boolean isRoot() {
-        return name.lastIndexOf(':') == name.length() - 1;
-    }
-
-    @Override
-    public String getScheme() {
-        return "nioMemLZF";
-    }
-
-}
-
-/**
- * This class represents an in-memory file.
- */
-class FileNioMem extends FileBase {
-
-    /**
-     * The file data.
-     */
-    FileNioMemData data;
-
-    private final boolean readOnly;
-    private long pos;
-
-    FileNioMem(FileNioMemData data, boolean readOnly) {
-        this.data = data;
-        this.readOnly = readOnly;
-    }
-
-    @Override
-    public long size() {
-        return data.length();
-    }
-
-    @Override
-    public FileChannel truncate(long newLength) throws IOException {
-        // compatibility with JDK FileChannel#truncate
-        if (readOnly) {
-            throw new NonWritableChannelException();
-        }
-        if (data == null) {
-            throw new ClosedChannelException();
-        }
-        if (newLength < size()) {
-            data.touch(readOnly);
-            pos = Math.min(pos, newLength);
-            data.truncate(newLength);
-        }
-        return this;
-    }
-
-    @Override
-    public FileChannel position(long newPos) {
-        this.pos = (int) newPos;
-        return this;
-    }
-
-    @Override
-    public int write(ByteBuffer src) throws IOException {
-        if (data == null) {
-            throw new ClosedChannelException();
-        }
-        int len = src.remaining();
-        if (len == 0) {
-            return 0;
-        }
-        data.touch(readOnly);
-        // offset is 0 because we start writing from src.position()
-        pos = data.readWrite(pos, src, 0, len, true);
-        src.position(src.position() + len);
-        return len;
-    }
-
-    @Override
-    public int read(ByteBuffer dst) throws IOException {
-        if (data == null) {
-            throw new ClosedChannelException();
-        }
-        int len = dst.remaining();
-        if (len == 0) {
-            return 0;
-        }
-        long newPos = data.readWrite(pos, dst, dst.position(), len, false);
-        len = (int) (newPos - pos);
-        if (len <= 0) {
-            return -1;
-        }
-        dst.position(dst.position() + len);
-        pos = newPos;
-        return len;
-    }
-
-    @Override
-    public int read(ByteBuffer dst, long position) throws IOException {
-        if (data == null) {
-            throw new ClosedChannelException();
-        }
-        int len = dst.remaining();
-        if (len == 0) {
-            return 0;
-        }
-        long newPos;
-        newPos = data.readWrite(position, dst, dst.position(), len, false);
-        len = (int) (newPos - position);
-        if (len <= 0) {
-            return -1;
-        }
-        dst.position(dst.position() + len);
-        return len;
-    }
-
-    @Override
-    public long position() {
-        return pos;
-    }
-
-    @Override
-    public void implCloseChannel() throws IOException {
-        pos = 0;
-        data = null;
-    }
-
-    @Override
-    public void force(boolean metaData) throws IOException {
-        // do nothing
-    }
-
-    @Override
-    public synchronized FileLock tryLock(long position, long size,
-            boolean shared) throws IOException {
-        if (data == null) {
-            throw new ClosedChannelException();
-        }
-        if (shared) {
-            if (!data.lockShared()) {
-                return null;
-            }
-        } else {
-            if (!data.lockExclusive()) {
-                return null;
-            }
-        }
-
-        return new FileLock(FakeFileChannel.INSTANCE, position, size, shared) {
-
-            @Override
-            public boolean isValid() {
-                return true;
-            }
-
-            @Override
-            public void release() throws IOException {
-                data.unlock();
-            }
-        };
-    }
-
-    @Override
-    public String toString() {
-        return data == null ? "<closed>" : data.getName();
-    }
-
-}
 
 /**
  * This class contains the data of an in-memory random access file.
@@ -794,5 +399,3 @@ class FileNioMemData {
     }
 
 }
-
-
