+++ h2/src/test/org/h2/test/db/TestOutOfMemory.java
@@ -17,8 +17,8 @@ import java.util.concurrent.atomic.AtomicReference;
 import org.h2.api.ErrorCode;
 import org.h2.mvstore.MVStore;
 import org.h2.store.fs.FilePath;
-import org.h2.store.fs.FilePathMem;
 import org.h2.store.fs.FileUtils;
+import org.h2.store.fs.mem.FilePathMem;
 import org.h2.test.TestBase;
 import org.h2.test.TestDb;
 import org.h2.util.Utils;
