+++ h2/src/test/org/h2/test/todo/TestTempTableCrash.java
@@ -10,8 +10,7 @@ import java.sql.DriverManager;
 import java.sql.Statement;
 import java.util.Random;
 import java.util.concurrent.TimeUnit;
-
-import org.h2.store.fs.FilePathRec;
+import org.h2.store.fs.rec.FilePathRec;
 import org.h2.test.unit.TestReopen;
 import org.h2.tools.DeleteDbFiles;
