+++ h2/src/test/org/h2/test/unit/TestFileSystem.java
@@ -28,9 +28,9 @@ import org.h2.message.DbException;
 import org.h2.mvstore.DataUtils;
 import org.h2.mvstore.cache.FilePathCache;
 import org.h2.store.fs.FilePath;
-import org.h2.store.fs.FilePathEncrypt;
-import org.h2.store.fs.FilePathRec;
 import org.h2.store.fs.FileUtils;
+import org.h2.store.fs.encrypt.FilePathEncrypt;
+import org.h2.store.fs.rec.FilePathRec;
 import org.h2.test.TestBase;
 import org.h2.test.utils.AssertThrows;
 import org.h2.test.utils.FilePathDebug;
