+++ h2/src/test/org/h2/test/unit/TestReopen.java
@@ -16,9 +16,9 @@ import org.h2.engine.Constants;
 import org.h2.engine.Database;
 import org.h2.engine.Session;
 import org.h2.message.DbException;
-import org.h2.store.fs.FilePathRec;
 import org.h2.store.fs.FileUtils;
 import org.h2.store.fs.Recorder;
+import org.h2.store.fs.rec.FilePathRec;
 import org.h2.test.TestBase;
 import org.h2.tools.Recover;
 import org.h2.util.IOUtils;
